+++ aerosense/__init__.py
@@ -13,8 +13,8 @@ Usage:
 from typing import List
 
 # --- Versioning ---
-__version__: str = "v4.2.1"
-__release__: str = "2026-03-23"
+__version__: str = "v4.3.0"
+__release__: str = "2026-03-24"
 
 # --- Component Exports ---
 from .core.controller import Controller